+++ src/main/groovy/ru/vyarus/gradle/plugin/lib/JavaLibExtension.groovy
@@ -1,0 +1,25 @@
+package ru.vyarus.gradle.plugin.lib
+
+import groovy.transform.CompileStatic
+
+/**
+ * Java-lib plugin extension for fine-tuning. Accessible as `javaLib` closure.
+ *
+ * @author Vyacheslav Rusakov
+ * @since 06.06.2021
+ */
+@CompileStatic
+class JavaLibExtension {
+
+    /**
+     * Used with java-platform plugin when platform is declared in the root module to rename artifact (which is by
+     * default equal to project name)
+     */
+    String bomArtifactId
+
+    /**
+     * Used with java-platform plugin when platform is declared in the root module to specify custom description in
+     * the generated pom (otherwise it would be root project description).
+     */
+    String bomDescription
+}