+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * Updated pom plugin (2.2.0): java-platform compatibility
+* Add java-platform plugin support (assuming declared platform published as BOM):
+    register pom plugin, configure 'bom' publication, add install task
 
 ### 2.1.0 (2020-01-19)
 * Updated pom plugin (2.1.0):